Use AppCache as a single source of truth
